Add files via upload
Copy my Odin Recipes project from the online course "The Odin Project" to use for my personal development and practice outside of the assigned manipulation of the files. This repository was created with a style.css file that was modified at the "Block and Inline" section of the course.  The recipe pages were modified with a new style.css file in their folder to allow different settings as needed on the recipe pages.